+++ frontend/src/main.jsx
@@ -4,18 +4,21 @@ import ReactDOM from 'react-dom/client';
 import { Notifications } from '@mantine/notifications';
 import { createBrowserRouter, RouterProvider } from 'react-router-dom';
 import MainLayout from './layouts/MainLayout';
-import Login from './views/Login';
 import indicesTheme from "./CustomProvider";
+
+import Login from './views/Login';
 import Inicio from './views/Inicio';
 import MiPerfil from "./views/MiPerfil";
+
 import SubirArchivos from "./views/SubirArchivos";
-import RegistroCarreras from "./views/RegCarreras";
-import RegistroPlanes from "./views/RegPlanes";
-import RegistroDiscapacidades from "./views/RegDiscapacidades";
-import IndicePermanencia from './views/IndPermanencia';
-import IndiceEgreso from "./views/IndEgreso";
-import IndiceDesercion from "./views/IndDesercion";
-import IndiceTitulacion from "./views/IndTitulacion";
+
+import RegistroCarreras from './views/registros/Carreras';
+import RegistroPlanes from './views/registros/Planes';
+import RegistroDiscapacidades from './views/registros/Discapacidades';
+import IndicePermanencia from './views/indices/Permanencia';
+import IndiceEgreso from './views/indices/Egreso';
+import IndiceDesercion from './views/indices/Desercion';
+import IndiceTitulacion from './views/indices/Titulacion';
 
 const router = createBrowserRouter([
   {
