+++ frontend/src/views/indices/Permanencia.jsx
@@ -1,43 +1,43 @@
 import { Flex, Group } from '@mantine/core';
-import Tabla from '../components/Tabla';
-import Header from '../components/header';
-import Dropdown from '../components/Dropdown';
+import Header from './../../components/header';
+import Tabla from './../../components/Tabla';
+import Dropdown from './../../components/Dropdown';
 
-const IndiceEgreso = () => {
+const IndicePermanencia = () => {
     // Informacion de prueba, no representa el comportamiento real
     const tabla = [
-        ['Semestre 1', '2015-1', '45','-','0.00%'],
-        ['Semestre 2', '2015-1', '45','-','0.00%'],
-        ['Semestre 4', '2015-1', '45','-','0.00%'],
-        ['Semestre 5', '2015-1', '45','-','0.00%'],
-        ['Semestre 6', '2015-1', '45','-','0.00%'],
-        ['','Acumulado', '225','']
+        ['Semestre 1', '2015-1', '45','-','15','30','66.75'],
+        ['Semestre 2', '2015-1', '45','-','15','30','66.75'],
+        ['Semestre 4', '2015-1', '45','-','15','30','66.75'],
+        ['Semestre 5', '2015-1', '45','-','15','30','66.75'],
+        ['Semestre 6', '2015-1', '45','-','15','30','66.75'],
+        ['','Acumulado', '','14','31']
     ];
 
     const headers = [
         ['Indices de rendimiento escolar cohorte generacional 2015-1 ingenieria mecanica'],
-        ['Semestre', 'Periodo', 'Activos', 'Egresados', 'Egreso terminal'],
+        ['Semestre', 'Periodo', 'Activos', 'Egresados','Desercion', 'Matricula final','Tasa de retencion'],
      ];
     return(
         <div style={{
             width: '100vw',
             padding: '3vw',
         }}>
-            <Header color="naranja" section="Indices" title="Egreso por cohorte generacional" route="/" />
+            <Header color="toronja" section="Indices" title="Permanencia por cohorte generacional" route="/" />
             <Flex direction="column">
                 <Group mt={0} mb={16}>
-                    <Dropdown  label="Programa educativo" color="#FFAA5A" data={[
+                    <Dropdown  label="Programa educativo" color="#FF785A" data={[
                         ['ISIC','Sistemas computacionales'],
                         ['QUI','Quimica'],
                         ['IND','Industrial'],
                     ]} />
-                    <Dropdown  label="Cohorte generacional" color="#FFAA5A" data={[
+                    <Dropdown  label="Cohorte generacional" color="#FF785A" data={[
                         ['2015-1','2015-2'],
                         ['2016-1','2016-1'],
                         ['2016-2','2016-2'],
                         ['2017-1','2017-1'],
                     ]} />
-                    <Dropdown  label="Cálculo de semestres" color="#FFAA5A" data={[
+                    <Dropdown  label="Cálculo de semestres" color="#FF785A" data={[
                         ['9','9 semestres'],
                         ['10','10 semestres'],
                         ['11','11 semestres'],
@@ -46,15 +46,15 @@ const IndiceEgreso = () => {
                         ['14','14 semestres'],
                         ['15','15 semestres'],
                     ]} />
-                    <Dropdown  label="Exportar" color="#FFAA5A" data={[
+                    <Dropdown  label="Exportar" color="#FF785A" data={[
                         ['Excel','Excel'],
                         ['PDF','PDF'],
                     ]} />
                 </Group>
-                <Tabla doubleHeader colors="tabla-naranja"  headers={headers} content={tabla} />
+                <Tabla doubleHeader colors="tabla-toronja"  headers={headers} content={tabla} />
             </Flex>
         </div>
     );
 };
 
-export default IndiceEgreso;
+export default IndicePermanencia;
