+++ frontend/src/views/registros/Discapacidades.jsx
@@ -1,8 +1,7 @@
 import { Button, Center, Flex, Group, TextInput, Textarea } from "@mantine/core";
 import { CirclePlus } from 'tabler-icons-react';
-import Header from '../components/header';
-import Tabla from "../components/Tabla";
-import './Registro.css';
+import Header from './../../components/header';
+import Tabla from './../../components/Tabla';
 
 const RegistroDiscapacidades = () => {
     const tabla = [
@@ -21,7 +20,7 @@ const RegistroDiscapacidades = () => {
         }}>
             <Header color="toronja" section="Registro" title="Discapacidades" route="/"/>
             <Group align="flex-start" spacing="3vw">
-                <Flex direction="column" className="col-input">
+                <Flex direction="column">
                     <form>
                         <TextInput label="Nombre" withAsterisk/>
                         <Textarea withAsterisk label="Descripción" autosize/>
